requirements.txt
flask
flask-cors
psutil
google-generativeai
Pillow
